Support startTime, user locale, and add new fields
Add new fields:
 - floors
 - minutesToFallAsleep
 - sleep StartTime

Adds the user's locale to the request, which is now used by fitbit to convert the units to the user's locale. See: https://dev.fitbit.com/docs/basics/#units

Fix a bug where the date was not using a zero-hour causing it never to match across rows.